pass small struct

+++ samples/test_c.c
@@ -1,12 +1,13 @@
 struct Point2D {
-    double x;
-    double y;
+    int x;
+    int y;
 };
 
-double f(struct Point2D xy)
+int f(struct Point2D xy)
 {
     return xy.x;
 }
+
 // struct Point2D f()
 // {
 //     struct Point2D point = { 10.0, 20.0 };
